chore: nettoyage client only
