+++ src/client/components/_commons/MiseEnPage/EnTete/EnTete.tsx
@@ -11,6 +11,27 @@ import { BoutonSeConnecter } from "@/components/_commons/BoutonSeConnecter";
 import { BoutonApplicationsPilote } from "@/components/_commons/MiseEnPage/EnTete/BoutonApplicationsPilote";
 import { ClientOnly } from "@/components/shared/ClientOnly";
 
+const InformationsEspaceConnecte = () => {
+  const { data: session } = useSession();
+
+  const peutVoirLeBoutonApplicationsPilote =
+    process.env.NEXT_PUBLIC_FF_PILOTE_EVAL === "true" &&
+    (process.env.NEXT_PUBLIC_FF_ACCES_PILOTE === "true" ||
+      session?.profil === "DITP_ADMIN");
+
+  if (session?.user != null)
+    return (
+      <>
+        {peutVoirLeBoutonApplicationsPilote ? (
+          <BoutonApplicationsPilote />
+        ) : null}
+        <Utilisateur />
+      </>
+    );
+
+  return <BoutonSeConnecter />;
+};
+
 const useEntete = () => {
   const { data: messageInformation } =
     api.gestionContenu.récupérerMessageInformation.useQuery();
@@ -28,10 +49,6 @@ export const EnTete = () => {
     messageInformation?.bandeauTexte ||
     "Des opérations de maintenance sont en cours et peuvent perturber le fonctionnement normal de PILOTE. En cas de difficultés : pilote.ditp@modernisation.gouv.fr";
   const bandeauType = messageInformation?.bandeauType || "WARNING";
-  const peutVoirLeBoutonApplicationsPilote =
-    process.env.NEXT_PUBLIC_FF_PILOTE_EVAL === "true" &&
-    (process.env.NEXT_PUBLIC_FF_ACCES_PILOTE === "true" ||
-      session?.profil === "DITP_ADMIN");
 
   return (
     <header className="fr-header" role="banner">
@@ -70,27 +87,22 @@ export const EnTete = () => {
             </div>
             <div className="fr-header__tools">
               <div className="fr-header__tools-links">
-                <ClientOnly>
-                  <div className="flex align-center gap-4">
-                    <BoutonContacterEquipePilote />
-                    {session?.user?.email ? (
-                      <>
-                        {peutVoirLeBoutonApplicationsPilote ? (
-                          <BoutonApplicationsPilote />
-                        ) : null}
-                        <Utilisateur email={session.user.email} />
-                      </>
-                    ) : (
-                      <BoutonSeConnecter />
-                    )}
-                  </div>
-                </ClientOnly>
+                <div className="flex align-center gap-4">
+                  <BoutonContacterEquipePilote />
+                  <ClientOnly>
+                    <InformationsEspaceConnecte />
+                  </ClientOnly>
+                </div>
               </div>
             </div>
           </div>
         </div>
       </div>
-      {session?.user ? <Navigation /> : null}
+      {session?.user ? (
+        <ClientOnly>
+          <Navigation />
+        </ClientOnly>
+      ) : null}
       {isBandeauActif ? (
         <BandeauInformation bandeauType={bandeauType}>
           {bandeauTexte}
