+++ src/client/components/_commons/MiseEnPage/EnTete/Utilisateur/Utilisateur.tsx
@@ -8,14 +8,10 @@ import { clsxm } from "@/utils/clsxm";
 import { Dropdown } from "@/components/shared/Dropdown";
 import api from "@/server/infrastructure/api/trpc/api";
 
-export const Utilisateur = ({ email }: { email: string }) => {
+export const Utilisateur = () => {
   const [estDeplie, setEstDeplie] = useState<boolean>(false);
-  // const { data: user } = api.profilUtilisateur.getUtilisateurConnecte.useQuery(
-  //   undefined,
-  //   {
-  //     staleTime: 5 * 60 * 1000,
-  //   },
-  // );
+  const [{ email }] =
+    api.profilUtilisateur.getUtilisateurConnecte.useSuspenseQuery();
 
   const { data: panelAdminEstDisponible } =
     api.gestionContenu.récupérerVariableContenu.useQuery({
